MUS-2479 Add end-to-end tests (#7)
This PR adds end-to-end tests for all of the example projects. My goal
with this PR is to pin the prod behavior of interacting with Open Cloud.
The tests make deployments to the Creator Store and then verify that
those changes were successfully deployed

This will also help with #6 to ensure that such a big change to the
upload flow won't cause any major regressions

A couple of the e2e tests fail right now but I'd like to move forward
regardless and will just leave the step as non-blocking. The bulk of the
checks should provide value for the other PRs I have in the pipeline